Create HomeButtons component and add work experience button
