+++ src/components/sections/home/Home.js
@@ -1,10 +1,10 @@
 import ThreeJSModel from "../../ThreeJS/ThreeJSModel";
 import AnimationName from "./animationName/AnimationName";
-import { Container, Typography, Button } from "@mui/material";
+import { Container, Typography } from "@mui/material";
 import { motion } from "framer-motion";
 import Icons from "../../icons/Icons";
-import { Link } from "react-scroll";
 import { useStyles } from "./Home-style";
+import HomeButtons from "./homeButtons/HomeButtons";
 
 const textVariants = {
   hidden: {
@@ -20,15 +20,7 @@ const textVariants = {
     },
   },
 };
-const buttonVariants = {
-  hidden: {
-    scale: 0,
-  },
-  visible: {
-    scale: 1,
-    transition: { type: "spring", stiffness: 100, delay: 2, duration: 3 },
-  },
-};
+
 const iconsVariants = {
   hidden: {
     opacity: 0,
@@ -62,26 +54,7 @@ const Home = (props) => {
         >
           Full-Stack Web Developer
         </Typography>
-        <motion.div className={classes.homeButtonsWrapper}>
-          <Button
-            variant="contained"
-            className={classes.homeButton}
-            component={motion.div}
-            variants={buttonVariants}
-            initial="hidden"
-            animate="visible"
-          >
-            <Link
-              to="projects"
-              spy={true}
-              smooth={true}
-              offset={-50}
-              duration={500}
-            >
-              My Projects
-            </Link>
-          </Button>
-        </motion.div>
+        <HomeButtons />
         <motion.div
           className={classes.icons}
           variants={iconsVariants}
